+++ backend/sivujetti/tests/bootstrap.php
@@ -2,6 +2,6 @@
 
 $doBootstrap = require __DIR__ . "/do-bootstrap.php";
 $doBootstrap(
-    // Use default site path (__DIR__ . "/test-site/")
-    // Use default plugins path (SIVUJETTI_BACKEND_PATH . "plugins/")
+    testSitePath: __DIR__ . "/test-site/",
+    testPluginsPath: __DIR__ . "/test-plugins/"
 );